+++ changelog.md
@@ -3,6 +3,8 @@ v0.3.2
 
 - Improved unit tests for Ohm's Law calculator.
 - Converted CalcVars object from Dictionary to List, closes #23.
+- Removed input variable to validator function, #25.
+- Added correct validators to IPC2221A track current calculator, closes #24.
 
 v0.3.1
 ------